+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - **BC break**: Automatically quote identifiers used in data parameters for
   `insert()` and `update()`. Implementations must remove any manual identifer 
   quotes that they were previously forced to include manually.
+### Removed
+- Previous deprecated *bind* parameter for `select()`, `update()` and
+  `delete()`, and passing a string to the *where* parameter.
 
 ## [1.2.0] - 2021-08-01
 ### Added